feature(add endpoint): added flexible microservice using FastAPI
for this endpoint we can use data models for detect the
correct format of data and integrity of this call.

added the requirements updated files for each instance and pipelines
for AWS sagemaker.

and the FastAPI endpoint was dockerized for external Use as an independ
and local sistem for test models without instances in cloud.